pms/requirements2.txt
aiohttp
aiosignal
annotated-types==0.6.0
anyio==3.7.1
argcomplete
astor==0.8.1
attrdict
attrs==21.4.0
Babel
bce-python-sdk
beautifulsoup4==4.8.2
blinker==1.7.0
boto3==1.26.12
botocore
cachetools==5.3.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer
click==8.1.7
compressed_rtf==1.0.6
contourpy==1.2.0
cssselect==1.2.0
cssutils==2.9.0
cycler==0.12.1
Cython==3.0.9
dataclasses-json==0.6.4
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
docx2txt==0.8
ebcdic==1.1.1
et-xmlfile==1.1.0
#extract-msg

fire
#Flask==2.2.2
flask-babel==3.0.0
Flask-Cors==4.0.0
fonttools==4.49.0
frozenlist==1.4.1
future==1.0.0
greenlet==3.0.3
groq==0.4.2
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
imageio==2.34.0
IMAPClient
imgaug
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.28
langchain-openai==0.0.8
langsmith==0.1.10
lazy_loader==0.3
lmdb==1.4.1
loguru==0.7.2
lxml==5.1.0
MarkupSafe==2.1.5
marshmallow==3.21.0
matplotlib==3.8.3
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
olefile==0.47
openai==1.13.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.15
packaging==23.2
paddleocr==2.7.0.3
paddlepaddle==2.6.0
pandas==2.2.1
pdf2docx==0.5.8
pdf2image==1.17.0
pdfminer.six
pillow==10.2.0
pinecone-client==2.2.2
plotly==5.19.0
premailer==3.10.0
protobuf==4.25.3
psutil==5.9.8
pyclipper==1.3.0.post5
pycryptodome==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.4.0
PyMuPDF==1.20.2
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil
python-docx==1.1.0
python-dotenv==1.0.1
python-pptx==0.6.23
pytz==2022.7
PyYAML==6.0.1
rapidfuzz==3.6.2
rarfile==4.1
regex==2023.12.25
requests==2.31.0
s3transfer
schedule==1.2.1
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
shapely==2.0.3
six
# ==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.27
tenacity==8.2.3
termcolor==2.4.0
textract
threadpoolctl==3.3.0
tifffile==2024.2.12
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.12
visualdl
#Werkzeug==3.0.1
wget==3.2
xlrd==1.2.0
XlsxWriter==3.2.0
yarl==1.9.4